+++ src/components/ProjectCarousel/ProjectCarousel.css
@@ -3,6 +3,7 @@
   border-radius: 0%;
   width: 10px;
   height: 10px;
+  outline: none;
 }
 
 .carousel .carousel-status {
